+++ src/main/java/com/modernhome/controller/DefectiveController.java
@@ -1,0 +1,38 @@
+package com.modernhome.controller;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping(value = "/production/defective")
+public class DefectiveController {
+	
+	private static final Logger logger = LoggerFactory.getLogger(DefectiveController.class);
+	
+	// ===========================================
+	
+	
+	// 의존성 주입
+	
+	
+	// ===========================================
+	
+	
+	// http://localhost:8088/production/defective/list
+	// 불량 목록 출력 (+ 검색)
+	
+	
+	// ===========================================
+	
+	
+	// 불량 등록 & 수정
+	
+	
+	// ===========================================
+	
+	
+	// 불량 삭제	
+	
+}
