+++ src/main/java/com/modernhome/persistence/PerformanceDAO.java
@@ -1,10 +1,10 @@
-package com.modernhome.service;
+package com.modernhome.persistence;
 
 import java.util.List;
 
 import com.modernhome.domain.WijoinVO;
 
-public interface WorkPrfrmService {
+public interface PerformanceDAO {
 	
 	// 작업지시 목록 중 품질검사가 완료인 목록
 	public List<WijoinVO> getqcEndInstr() throws Exception;
