+++ src/main/resources/mappers/defectiveMapper.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.modernhome.mapper.DefectiveMapper">
+	
+	<!-- Join 결과 매핑 -->
+	<resultMap id="wijoinRM" type="WijoinVO">
+		
+	</resultMap>
+	<!-- Join 결과 매핑 -->
+
+</mapper>
